Attempt to refactor the peptides_df fixture to fix the ci pytest
