+++ tests/protzilla/data_preprocessing/test_filter_samples.py
@@ -10,7 +10,6 @@ from protzilla.data_preprocessing.filter_samples import (
     by_proteins_missing_plot,
 )
 from tests.protzilla.data_preprocessing.test_peptide_preprocessing import (
-    peptides_df,
     assert_peptide_filtering_matches_protein_filtering,
 )
 
@@ -52,11 +51,23 @@ def filter_samples_df():
 
 
 def test_by_proteins_missing(filter_samples_df, show_figures, peptides_df):
-    method_input1 = {"protein_df": filter_samples_df, "peptide_df": None, "percentage": 0.5}
+    method_input1 = {
+        "protein_df": filter_samples_df,
+        "peptide_df": None,
+        "percentage": 0.5,
+    }
     method_output1 = by_proteins_missing(**method_input1)
-    method_input2 = {"protein_df": filter_samples_df, "peptide_df": peptides_df, "percentage": 0.6}
+    method_input2 = {
+        "protein_df": filter_samples_df,
+        "peptide_df": peptides_df,
+        "percentage": 0.6,
+    }
     method_output2 = by_proteins_missing(**method_input2)
-    method_input3 = {"protein_df": filter_samples_df, "peptide_df": None, "percentage": 0.65}
+    method_input3 = {
+        "protein_df": filter_samples_df,
+        "peptide_df": None,
+        "percentage": 0.65,
+    }
     method_output3 = by_proteins_missing(**method_input3)
 
     list_samples_excluded_1 = method_output1["filtered_samples"]
@@ -98,10 +109,19 @@ def test_by_proteins_missing(filter_samples_df, show_figures, peptides_df):
         "Sample",
     )
 
+
 def test_filter_samples_by_protein_count(filter_samples_df, show_figures, peptides_df):
-    method_input1 = {"protein_df": filter_samples_df, "peptide_df": None, "deviation_threshold": 0.3}
+    method_input1 = {
+        "protein_df": filter_samples_df,
+        "peptide_df": None,
+        "deviation_threshold": 0.3,
+    }
     method_output1 = by_protein_count(**method_input1)
-    method_input2 = {"protein_df": filter_samples_df, "peptide_df": peptides_df, "deviation_threshold": 1.0}
+    method_input2 = {
+        "protein_df": filter_samples_df,
+        "peptide_df": peptides_df,
+        "deviation_threshold": 1.0,
+    }
     method_output2 = by_protein_count(**method_input2)
 
     list_samples_excluded_1 = method_output1["filtered_samples"]
@@ -139,10 +159,21 @@ def test_filter_samples_by_protein_count(filter_samples_df, show_figures, peptid
         "Sample",
     )
 
-def test_filter_samples_by_protein_intensity_sum(filter_samples_df, show_figures, peptides_df):
-    method_input1 = {"protein_df": filter_samples_df, "peptide_df": None, "deviation_threshold": 1.0}
+
+def test_filter_samples_by_protein_intensity_sum(
+    filter_samples_df, show_figures, peptides_df
+):
+    method_input1 = {
+        "protein_df": filter_samples_df,
+        "peptide_df": None,
+        "deviation_threshold": 1.0,
+    }
     method_output1 = by_protein_intensity_sum(**method_input1)
-    method_input2 = {"protein_df": filter_samples_df, "peptide_df": peptides_df, "deviation_threshold": 0.3}
+    method_input2 = {
+        "protein_df": filter_samples_df,
+        "peptide_df": peptides_df,
+        "deviation_threshold": 0.3,
+    }
     method_output2 = by_protein_intensity_sum(**method_input2)
 
     list_samples_excluded_1 = method_output1["filtered_samples"]
